# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Website API Documentation",
        "description": "API documentation for AI Website backend service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vendors/{vendor}/wallet": {
            "get": {
                "tags": [
                    "Vendor - Wallet"
                ],
                "summary": "Get vendor wallet overview and recent transactions",
                "operationId": "b61122f24e68e16feb0c68e4d6034233",
                "parameters": [
                    {
                        "name": "vendor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendor wallet details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/App\\Http\\Resources\\VendorWalletResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/vendors/{vendor}/payouts": {
            "post": {
                "tags": [
                    "Vendor - Wallet"
                ],
                "summary": "Submit payout withdrawal request",
                "operationId": "fb56128e68a46aa843ebe37c73c1e972",
                "parameters": [
                    {
                        "name": "vendor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App\\Http\\Requests\\Vendor\\CreatePayoutRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payout request submitted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/App\\Http\\Resources\\PayoutRequestResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}